% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_function.R
\name{condExpectation}
\alias{condExpectation}
\title{Conditional expectation for the SIR method.}
\usage{
condExpectation(data, y, H, scale = TRUE)
}
\arguments{
\item{data}{List; The list of tables to study.}

\item{y}{A numerical variable of interest.}

\item{H}{Integer; The number of slices to cut y.}

\item{scale}{Logicial; Should the data be scaled by the Frobenius norm of the
covariance matrix (TRUE) or not (FALSE)? Default is TRUE.}
}
\value{
a list consisting of
\itemize{
 \item{pdata}{ the conditional expectations of X given y (list of lenght T)}
 \item{weights}{ the number of observations in every slides}
 }
}
\description{
\code{condExpectation} computes the conditional expectation of X given y with
a slicing approach.
}

