% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interstructure.R
\name{interStructure}
\alias{interStructure}
\title{Inter-structure analysis.}
\usage{
interStructure(data, method = c("sir", "classic", "dual"), y = NULL,
  H = NULL, scale = TRUE)
}
\arguments{
\item{data}{List; The list of tables to study.}

\item{method}{The method to apply ("sir", "classic" or "dual"). Default is
"sir".}

\item{y}{A numerical variable of interest (mandatory for "sir").}

\item{H}{Integer; The number of slices to cut y (mandatory for "sir").}

\item{scale}{Logical; Should the interstructure be computed with the cosine
(TRUE) or with the dot product (FALSE). Default is TRUE.}
}
\value{
S3 \code{interstructure} object; a list consisting of
\itemize{
 \item{objects}{ representative objects according to the chosen method}
 \item{corr}{ matrix used for the eigendecomposition}
 \item{values}{ eigenvalues of the decomposition of the RV coefficient
 or correlation coefficient matrix}
 \item{vectors}{ eigenvectors of the decomposition of the RV coefficients
 or correlation coefficient matrix}
 \item{consensus}{ interstructure consensual matrix}
 \item{tweights}{ computed weights for each table}
 \item{pdata}{ a list with two elements: \code{pdata} the preprocessed data
 (list with length T) used to compute the scalar products (centered and
 eventually scaled; conditional expectations for "sir" as produced by
 \link{condExpectation}) and \code{weights} the weights associated with every
  observation.}
 }
}
\description{
\code{interStructure} perfoms the inter-structure analysis and returns its
result (table weights, compromise similarity matrix between tables,...). The
function is also called by \code{\link{statis}}. You should only use it when
wanting to perform only the inter-structure analysis.
}

