% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statis.R
\name{statis}
\alias{statis}
\title{STATIS, dual STATIS and SIR-STATIS for cubic data.}
\usage{
statis(data, method = c("sir", "classic", "dual"), y = NULL, H = NULL,
  scale = TRUE)
}
\arguments{
\item{data}{List; The list of tables to study.}

\item{method}{The method to apply ("sir", "classic" or "dual"). Default is
"sir".}

\item{y}{A numerical variable of interest (mandatory for "sir").}

\item{H}{Integer; The number of slices to cut y (mandatory for "sir").}

\item{scale}{Logical; Should the interstructure be computed with the cosine
(TRUE) or with the dot product (FALSE). Default is TRUE.}
}
\value{
S3 object of class
\code{statisRes}: a list consisting of
\itemize{
   \item{parameters}{
   \itemize{
     \item{method}{ method used ("sir", "classic" or "dual")}
     \item{H}{ number of slices}
     \item{scale}{ logical indicating if the interstructure is analyzed with
     the cosine (TRUE) or the dot product (FALSE)}
     }
   }
   \item{svd}{
   \itemize{
     \item{values}{ eigenvalues of the generalized svd (square of singular
     values)}
     \item{P}{ left-singular vectors of the generalized svd}
     \item{Q}{ right-singular vectors of the generalized svd}
     }
   }
   \item{observations}{
   \itemize{
     \item{coord}{ if \code{method} is "sir" or "dual", matrix with the
     compromise coordinates of slices ("sir") or of observations (center of
     gravity of the table-specific coordinates). If \code{method} is
     "classic", matrix with the coordinates of the observations (factors).}
     \item{tcoord}{ if \code{method} is "sir" or "dual", list of matrices
     with the table-specific coordinates of slices ("sir") or of
     observations. If \code{method} is "classic", list of the coordinates of
     the partial (table-specific) factors}
     }
   }
   \item{variables}{
   \itemize{
     \item{cos2}{ if \code{method} is "sir" or "dual", matrix with the
     coordinates of the variables on the circle of correlation (also called
     squared cosinus). If \code{method} is "classic", matrix with the
     compromise coordinates of the variables on the circle of correlation
     (center of gravity of the table-specific coordinates)}
     \item{tcos2}{ if \code{method} is "sir" or "dual", list of matrices with
     the coordinates of the partial (table-specific) squared-cosinus. If
     \code{method} is "classic", list of the table-specific coordinates}
     }
   }
 }
 @export
}
\description{
\code{statis} performs the STATIS-like analyses of cubic data with the two
standard steps: inter-structure analysis (with the function
\code{\link{interStructure}}) and intra-structure analysis.
}
\details{
STATIS, dual STATIS and SIR-STATIS for cubic data

perform STATIS-like analyses of cubic data

The package is well suited to explore time-course datasets, eventually in
relation with a target numeric variable (SIR approach).
}

